MIgrate page to route folder
