+++ web/src/App.jsx
@@ -1,8 +1,8 @@
 import { Routes, Route } from "react-router-dom";
-import HomeRoute from "./routes/HomeRoute.component";
-import { AuthenticationRoute } from "./routes/AuthenticationRoute.component";
-import NavigationRoute from "./routes/NavigationRoute.component";
-import ShopRoute from "./routes/ShopRoute.component";
+import HomeRoute from "./routes/home/HomeRoute.component";
+import AuthenticationRoute from "./routes/authentication/AuthenticationRoute.component";
+import NavigationRoute from "./routes/navigation/NavigationRoute.component";
+import ShopRoute from "./routes/shop/ShopRoute.component";
 import CheckoutRoute from './routes/checkout/CheckoutRoute.component';
 
 function App() {
