--- web/src/layouts/DashboardLayout.component.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Dashboard() {
-  return (
-    <div>Dashboard</div>
-  )
-}
-
-export default Dashboard;
