--- web/src/routes/AuthenticationRoute.component.jsx
@@ -1,11 +1,0 @@
-import AuthenticationPage from "../pages/authentication/AuthenticationPage.component";
-
-function AuthenticationRoute() {
-  return (
-    <div>
-      <AuthenticationPage />
-    </div>
-  );
-}
-
-export { AuthenticationRoute };
