--- web/src/routes/DashboardRoute.component.jsx
@@ -1,7 +1,0 @@
-function DashboardRoute() {
-  return (
-    <div>DashboardRoute</div>
-  )
-}
-
-export default DashboardRoute;
