--- web/src/routes/HomeRoute.component.jsx
@@ -1,17 +1,0 @@
-import { useEffect, useState } from 'react';
-import CategoriesContainer from '../components/categories/CategoriesContainer.component';
-
-function HomeRoute() {
-  const [categories, setCategories] = useState([]);
-
-  useEffect(() => {
-    const categories = require('../data/categories.json');
-    setCategories(categories);
-  }, [])
-
-  return (
-    <CategoriesContainer categories={categories} />
-  )
-}
-
-export default HomeRoute;
