--- web/src/routes/NavigationRoute.component.jsx
@@ -1,9 +1,0 @@
-import HomeLayout from '../layouts/HomeLayout.component';
-
-function NavigationRoute() {
-  return (
-    <HomeLayout />
-  )
-}
-
-export default NavigationRoute;
