--- web/src/routes/ShopRoute.component.jsx
@@ -1,9 +1,0 @@
-import Shop from '../pages/shop/Shop.component';
-
-function ShopRoute() {
-  return (
-    <div><Shop /></div>
-  )
-}
-
-export default ShopRoute;
