+++ web/src/routes/authentication/AuthenticationRoute.component.jsx
@@ -1,0 +1,15 @@
+import SignInForm from "../../components/signin-form/SignInForm.component";
+import SignUpForm from "../../components/signup-form/SignUpForm.component";
+
+function AuthenticationRoute() {
+  return (
+    <div>
+      <div className="authentication-container">
+        <SignInForm />
+        <SignUpForm />
+      </div>
+    </div>
+  );
+}
+
+export default AuthenticationRoute;
