+++ web/src/routes/shop/shop.style.scss
